Code refactoring and updated readme.md
