+++ ui-library-filesharing-chat-composite/app/package.json
@@ -6,7 +6,7 @@
     "@azure/communication-calling": "^1.4.4",
     "@azure/communication-chat": "^1.1.0",
     "@azure/communication-react": "^1.2.2-alpha-202204230012.0",
-    "axios": "^0.27.1",
+    "axios": "^0.26.1",
     "form-data": "^4.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
