+++ ui-library-filesharing-chat-composite/app/src/index.css
@@ -1,10 +1,5 @@
-body {
-  margin: 0;
+html,
+body,
+#root {
   height: 100%;
-  width: 100%;
-  /* font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif; */
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
 }
